chore: minor fix
